+++ src/apis/dlog.py
@@ -460,39 +460,9 @@ async def dlogLeaderboard(request: Request, response: Response, authorization: s
 
     return {"error": False, "response": {"list": ret[(page - 1) * pagelimit : page * pagelimit], "page": str(page), "tot": str(tot)}}
 
-@app.get(f"/{config.vtc_abbr}/dlog/newdrivers")
-async def dlogNewDriver(request: Request, response: Response, authorization: str = Header(None), pagelimit: Optional[int] = 10):
-    rl = ratelimit(request.client.host, 'GET /dlog/newdrivers', 60, 60)
-    if rl > 0:
-        response.status_code = 429
-        return {"error": True, "descriptor": f"Rate limit: Wait {rl} seconds"}
-        
-    au = auth(authorization, request, allow_application_token = True)
-    if au["error"]:
-        response.status_code = 401
-        return au
-    
-    if pagelimit <= 1:
-        pagelimit = 1
-    elif pagelimit >= 250:
-        pagelimit = 250
-
-    conn = newconn()
-    cur = conn.cursor()
-
-    cur.execute(f"SELECT userid, joints FROM driver WHERE userid >= 0 ORDER BY joints DESC LIMIT {pagelimit}")
-    t = cur.fetchall()
-    ret = []
-    for tt in t:
-        cur.execute(f"SELECT name, discordid, avatar FROM user WHERE userid = {tt[0]}")
-        p = cur.fetchall()
-        ret.append({"userid": str(tt[0]), "name": p[0][0], "discordid": str(p[0][1]), "avatar": p[0][2], "joints": str(tt[1])})
-
-    return {"error": False, "response": {"list": ret, "page": "1", "tot": str(pagelimit)}}
-
 @app.get(f"/{config.vtc_abbr}/dlogs")
 async def dlogs(request: Request, response: Response, authorization: str = Header(None), \
-    page: Optional[int] = -1, speedlimit: Optional[int] = 0, quserid: Optional[int] = -1, \
+    page: Optional[int] = -1, order: Optional[str] = "desc", speedlimit: Optional[int] = 0, quserid: Optional[int] = -1, \
         starttime: Optional[int] = -1, endtime: Optional[int] = -1, game: Optional[int] = 0, pagelimit: Optional[int] = 10):
     rl = ratelimit(request.client.host, 'GET /dlogs', 60, 60)
     if rl > 0:
@@ -521,6 +491,10 @@ async def dlogs(request: Request, response: Response, authorization: str = Heade
     elif pagelimit >= 250:
         pagelimit = 250
 
+    if not order in ["asc", "desc"]:
+        order = "asc"
+    order = order.upper()
+
     limit = ""
     if quserid != -1:
         limit = f"AND userid = {quserid}"
@@ -538,7 +512,7 @@ async def dlogs(request: Request, response: Response, authorization: str = Heade
     if game == 1 or game == 2:
         gamelimit = f" AND unit = {game}"
 
-    cur.execute(f"SELECT userid, data, timestamp, logid, profit, unit, distance FROM dlog WHERE logid >= 0 {limit} {timelimit} {speedlimit} {gamelimit} ORDER BY timestamp DESC LIMIT {(page - 1) * pagelimit}, {pagelimit}")
+    cur.execute(f"SELECT userid, data, timestamp, logid, profit, unit, distance FROM dlog WHERE logid >= 0 {limit} {timelimit} {speedlimit} {gamelimit} ORDER BY logid {order} LIMIT {(page - 1) * pagelimit}, {pagelimit}")
     
     t = cur.fetchall()
     ret = []