+++ src/plugins/announcement.py
@@ -14,7 +14,7 @@ import multilang as ml
 
 @app.get(f"/{config.vtc_abbr}/announcements")
 async def getAnnouncement(request: Request, response: Response, authorization: str = Header(None), \
-    page: Optional[int]= -1, aid: Optional[int] = -1, pagelimit: Optional[int] = 10):
+    page: Optional[int]= -1, aid: Optional[int] = -1, order: Optional[str] = "desc", pagelimit: Optional[int] = 10):
     rl = ratelimit(request.client.host, 'GET /announcement', 60, 60)
     if rl > 0:
         response.status_code = 429
@@ -43,6 +43,10 @@ async def getAnnouncement(request: Request, response: Response, authorization: s
     elif pagelimit >= 100:
         pagelimit = 100
 
+    if not order in ["asc", "desc"]:
+        order = "asc"
+    order = order.upper()
+
     if aid != -1:
         cur.execute(f"SELECT title, content, atype, timestamp, userid, aid, pvt FROM announcement WHERE aid = {aid} {limit}")
         t = cur.fetchall()
@@ -61,7 +65,7 @@ async def getAnnouncement(request: Request, response: Response, authorization: s
     if page <= 0:
         page = 1
 
-    cur.execute(f"SELECT title, content, atype, timestamp, userid, aid, pvt FROM announcement WHERE aid >= 0 {limit} ORDER BY timestamp DESC LIMIT {(page-1) * pagelimit}, {pagelimit}")
+    cur.execute(f"SELECT title, content, atype, timestamp, userid, aid, pvt FROM announcement WHERE aid >= 0 {limit} ORDER BY aid {order} LIMIT {(page-1) * pagelimit}, {pagelimit}")
     t = cur.fetchall()
     ret = []
     for tt in t: