+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+**v1.10.7**  
+1.Removed **GET** `/dlog/newdrivers`  
+2.Removed `sort_by_highest_role` parameter from **GET** `/members`  
+3.Added `order`, `order_by` parameter for **GET** `/members`, **GET** `/users`  
+4.Added `order` parameter for **GET** `/dlogs`, **GET** `/applications`, **GET** `/announcements`  
+
 **v1.10.6**  
 1.Added support to email & password login  
 2.Added time-range query for `/dlog/stats`